% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cities}
\alias{cities}
\title{Data with São Paulo state cities and codes}
\format{A data frame with 645 rows and 2 variables:
\describe{
  \item{code}{City's code, from 1 to 645}
  \item{city}{City's name}
  ...
}}
\usage{
cities
}
\description{
This is data with São Paulo state cities and their respective codes
to use with the other functions
}
\keyword{datasets}
