% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_table.R
\name{download_table_sp}
\alias{download_table_sp}
\title{Download tables publicized by the São Paulo's Security Office}
\usage{
download_table_sp(ano, municipio, type)
}
\arguments{
\item{ano}{Year to download}

\item{municipio}{Intended city. Is a number between 1 and 645, representing the rank of the city in alphabetical order.}

\item{type}{Formulary to make the request, encodes the type of information to be downloaded.
"ctl00$conteudo$$btnMes" downloads recorded crime count.}
}
\description{
Download tables publicized by the São Paulo's Security Office
}
