% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.R
\name{get_detailed_table_sp}
\alias{get_detailed_table_sp}
\title{Download detailed information publicized by São Paulo's Security Office}
\usage{
get_detailed_table_sp(folder, year, month, department,
  url = "http://www.ssp.sp.gov.br/transparenciassp/", ...)
}
\arguments{
\item{year}{Year to download.}

\item{month}{Intended city. Is a number between 1 and 645, representing the rank of the city in alphabetical order.}

\item{type}{Formulary to make the request, encodes the type of information to be downloaded.}
}
\description{
Download detailed information publicized by São Paulo's Security Office
}
