% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.R
\name{get_summarized_table_sp}
\alias{get_summarized_table_sp}
\title{Download tables publicized by the São Paulo's Security Office}
\usage{
get_summarized_table_sp(year, city, type = "ctl00$conteudo$btnMensal")
}
\arguments{
\item{year}{Year to download.}

\item{city}{Intended city. Is a number between 1 and 645, representing the rank of the city in alphabetical order.}

\item{type}{Formulary to make the request, encodes the type of information to be downloaded.
"ctl00$conteudo$$btnMes" downloads recorded crime count.}
}
\description{
Download tables publicized by the São Paulo's Security Office
}
