% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{police_stations}
\alias{police_stations}
\title{Data with São Paulo police stations}
\format{A data frame with 1465 rows and 2 variables:
\describe{
  \item{code}{City's code, from 1 to 645}
  \item{police_station}{police stations' name}
  ...
}}
\usage{
police_stations
}
\description{
This is data with São Paulo state police stations to be used
with other functions
}
\keyword{datasets}
