% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.R
\name{get_summarized_table_sp}
\alias{get_summarized_table_sp}
\alias{get_historical_summarized_table_sp}
\title{Download tables publicized by the São Paulo's Security Office}
\usage{
get_summarized_table_sp(year = NULL, city = NULL,
  type = "ctl00$conteudo$btnMensal", region = "0", police_station = "0")

get_historical_summarized_table_sp()
}
\arguments{
\item{year}{Year to download.}

\item{city}{Intended city. Is a number between 1 and 645, representing the rank of the city in alphabetical order.}

\item{type}{Formulary to make the request, encodes the type of information to be downloaded.
"ctl00$conteudo$$btnMes" downloads recorded crime count.}

\item{region}{Intended region. Is a number between 0 and 9, representing an administrative region.}

\item{police_station}{number of the police station}
}
\description{
Download tables publicized by the São Paulo's Security Office

Downloads summarized crime tables from multiple years and cities
}
\section{Functions}{
\itemize{
\item \code{get_historical_summarized_table_sp}: Allows for getting summarized tables from
multiple years and cities.
}}

